requirements.txt
pillow
imagehash
opencv-python
pywin32
mss
pytesseract
rapidfuzz
PySide6